// src/interfaces/species_types.ts
export default interface Species {
        species_data: {
          "name": string
          "classification": string
          "designation": string
          "average_height": string
          "average_lifespan": string
          "hair_colors": string
          "skin_colors": string
          "eye_colors": string
          "homeworld": string
          "language": string
          "people": string[]
          "films": string[]
          "url": string
          "created": string
          "edited": string
        }
}